sync: Configuration of the execution
The sync operation is more an exception then a required
background task that is why we need to have ability to
configure if we want it to run and how ofter it should run.
The default for it is to run each 5 min. If min_interval is
set to 0 we won't start the task at all.
Also as each storage can have lot of repositories it makes
sense to process them in batches. The size of the batch is
another configuration option for the sync operation we want.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3719