objectpool: Convert `Link()` to write alternates transactionally
When linking a repository into an object pool, then we must write the
alternates file to point to the object pool's object storage such that
it reuses objects from it. This is currently done in a non-transactional
way and can thus easily race with concurrent updates of the same
gitconfig.

Convert it to use our new locking file writer and voting infrastructure
to make this update race-free.

Changelog: fixed