housekeeping: Convert `Perform()` to write config transactionally
When performing housekeeping tasks, one of these tasks is to clean up
old config entries. This is currently done in a non-transactional way
and can thus easily race with concurrent updates of the same gitconfig.

Convert it to use our new `localrepo.SetConfig()` function, which is
properly aware of transactional semantics.

Changelog: fixed